requirements.txt
amqp==2.5.0
arrow==0.14.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
base58==0.2.5
billiard==3.6.0.0
bitcash==0.5.2.8
boto3==1.15.12
botocore==1.18.12                 
cashaddress==1.0.4
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coincurve==12.0.0
coinex.py==1.0.4
cryptography==2.7
decorator==4.4.0
Django==2.2.1
django-constance==2.5.0
django-cors-headers==3.0.2
django-rest-hooks==1.5.0
djangorestframework==3.9.2
django-filter==2.2.0
django-picklefield==3.0.1
emoji==0.5.4
Fabric3==1.14.post1
future==0.17.1
gcloud==0.18.3
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
importlib-metadata==0.18
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
kombu==4.6.3
# meinheld==1.0.2
Markdown==3.1.1
meld3==1.0.2
mock==3.0.5
more-itertools==7.2.0
oauth2client==4.1.3
oauthlib==3.0.2
packaging==19.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.12.0
praw==6.3.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-mock==1.10.4
python-dateutil==2.8.0
python-decouple==3.1
python-telegram-bot==13.6
pytz==2019.1
random-username==1.0.2
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
requests-mock==1.6.0
rsa==4.0
six==1.12.0
sseclient==0.0.24
sqlparse==0.3.0
supervisor==4.2.0
traitlets==4.3.2
telegram==0.0.1
Telethon==1.6.2
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
web3==5.17.0
websocket-client==0.56.0
whitenoise==4.1.2
zipp==0.5.2
git+https://github.com/joemarct/python-twitter@1c2d58022150a2b0207563e0aed496d433e6a463
git+https://github.com/joemarct/pywallet@d26820975d4548a1867c1b96d52a3842c1d138c8